Stop merging by tags in add_reserved_diagnostics
Currently, add_reserved_diagnostics merges histograms based on storyTags,
which results in us having merged histograms for each tag set.

This is a problem because the dashboard does not and should not consider
story tags when computing test paths, so you end up with duplicates. Since
story names are unique irrespective of tags, we should not need to consider
story tags at all when merging. Fixes #4122.

Bug: catapult:#4122
Change-Id: If5d593b1143a72ed1d06e900ebe132478247c7b9