Purge the OS pagecache selectively on Android
Add clear_system_cache binaries for Android ARM 32/64 and use them to clear the
pagecache for the user data dir.

Removing the DEX/DSO from pagecache is less trivial because it requires finding
the path to the APK, which seems fragile and better be left for a separate
change.

Purging the APK/DEX/DSO from pagecache is probably not representative on N+.
This is because a lot of apps use WebView, which brings the same native library
to pagecache (with Monochrome). Since the majority of users is on M-, we should
still purge the pagecache for APK/DEX/DSO on the bots, which also provides 'perf
coverage' for the 'orderfile pipeline'.

(!!!) Note to perf sheriffs: this should move the startup metrics on Android
bots _massively_. This is an intentional attempt to make those metrics more
representative.

Bug: chromium:760498
Change-Id: I84bf23d6bf2d8af1081cc8b3ad14129a4b4ffb04