Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you add test for this?

The cause of the issue was a race that only showed up when running on the actual cros vm (and only in a particular scenario). So it could only really be caught by bots on cros waterfall, where it was in fact caught by testBrowserCreation.

I could mock out the the self._cri and and some unittests here under a couple of fake racy conditions, but not too sure if that would be really valuable?

Patch-set: 1
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0