[TBMv2] Switch to dynamically importing diagnostics in Python
Right now, since all diagnostic subclasses are contained in histogram.py, it's
enough for all_diagnostics to maintain a map from typenames to classes.
However, when we move diagnostic subclasses into their own files, we need to
avoid a circular import problem as follows:

* `diagnostic` must import `all_diagnostics` to look up classes for
  deserialization
* `all_diagnostics` must import all known diagnostic modules, to include them
  in the dict.
* Diagnostic subclasses must import `diagnostic` for subclassing.

To work around this, we change the map stored in all_diagnostics to map from
diagnostic typenames to the modules that contain them, so we can dynamically
import each module.

To assist users, we encapsulate this functionality in a new function
`all_diagnostics.GetDiagnosticClassForName(name)`, which can be used
equivalently of the old all_diagnostics.DIAGNOSTICS_BY_NAME.

This CL unblocks moving Diagnostic subclasses out of histogram.py.

Bug: catapult:#3613
Change-Id: Ib08bffc8a2b8d7523152227b9600943aa1056d35