[Telemetry] Factor out _GetDevToolsClient to chrome_browser_backend
All three subclasses (desktop, android and cros) now have essentially
the same implementation of _GetDevToolsClientConfig; we can thus
simplify the code by moving this up to the parent class.

Also, given that (1) getting the devtools config, and (2) checking
whether the agent is ready, have to be both retried together in case
of failure, we can further simplify on each individual backend the
code for _FindDevToolsPortAndTarget.

Bug: chromium:787834
Bug: catapult:#1977
Change-Id: I761d7c2ca658903258ff244db9bc3b6ff7a01feb