Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> This is an internal-only change, right?
> 
> I believe this line doesn't break, but just want to be sure:
> https://cs.corp.google.com/chromeos_public/src/third_party/autotest/files/client/common_lib/cros/chrome.py?l=157

Thanks for pointing that out! Yes, this is an internal only change, so should not affect things. On that file I see that you mutate the list of extensions_to_load *before* sending it to possible_browser.Create, which should be fine. In this CL the only change I make is freezing (a copy of) the list within the browser instance so we don't modify it any more after creation.

Patch-set: 2
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0