[Telemetry] Move android startup args to its possible-browser
Move the android-specific startup args computation out of the
browser-backend and into the possible-browser.

For this we make the profile_directory property accessible
from both the backend and the possible-browser.

Also do not cache the _profile_directory in the backend, there is no
real need since it's not an expensive operation, and this helps reduce
the amount of state stored on the backend.

Bug: chromium:787834
Change-Id: Iaaee5e915e201b894cc7b89402d1c576c38b408f