[Telemetry] Switch to possible_browser.Create(browser_options)
To create a browser only the browser_options (and not the full
finder_options) should be needed.

This CL changes the API so, momentarily, both finder_options and
browser_options may be passed to possible_browser.Create; and
migrates clients within catapult.

After external clients are also migrated, we can then remove the
check so only browser_options are allowed in the future.

Bug: chromium:801214
Change-Id: Ic254da51600c1ebc3be4cd316c251c284e514495