[Telemetry] New API to create browsers
This CL introduces thew new API methods to create a browser out of a
possible_browser. The old API still works.

This is needed to support use cases where setting up the browser
environment (preparing a profile directory, or writing a command line
file) needs to be separated from actually starting the browser.

TBR=achuith@chromium.org
Bug: chromium:801578
Change-Id: I0ae9aac2f5e44540b3b27e4c5b2777a301b54e60