[Telemetry] Fix bug when reading devtools port in desktop
The previous code had a bug which forgot to initialize `lines` in the
case where the devtools file is already present but with a size in
bytes of zero.

There was a small chance for this to happen, but it did often happen
and raised a UnboundLocalError due to the variable not being defined.

Bug: chromium:801667
Change-Id: I5e0d8786946cdc8ff69b67cb1287b91f56aab475