Emit output when trace2html is complete
We should be emitting output after all long-running subtasks are
complete in order to give hints to swarming that we're making forward
progress and therefore shouldn't run up against the I/O timeout. In this
case, it's not unusual for this command to take upwards of a minute and
it's immediately followed by another long-running task (calculating
metrics from the trace).

Bug: chromium:801596
Change-Id: Ia393ca6d5120cc3d7f5e46c82911fb82bca26204