Purge the OS pagecache for APK, libchrome.so
Also flush OS pagecache for everything that happens to be in a
(sub)directory from the base.apk, which is useful only for L+.

Avoid flushing on N+ because the case of libchrome/DEX/etc being paged
out seems non-typical.

(!!!) Note to perf sheriffs: this should move the startup metrics on
Android, which will likely cause regression alerts. This is an
intentional attempt to make those metrics more representative.

Bug: chromium:760498
Change-Id: Iae231d3c3253a257b22e8683166225efb229ca95