[Telemetry] Refactor startup tracing test
The testStartupTracing* tests need to create their own browser instance,
so they shouldn't reuse TabTestCase (or BrowserTestCase).

In the older implementation they closed the browser created for them by
the class without using it (thus preventing other tests in the class
from reusing it), and then went on to create a new one which wouldn't
be reusable.

The code is simpler (and the running time for tests probably faster) if
we just make this test responsible for managing the lifetime of its own
browser.

Also consolidate tests for all platforms in a single testStartupTracing
method.

Bug: chromium:801578
Change-Id: I59cc01994c091de8c4df8bb0656acb37bb4c50c6