[Dashboard] Don't create extra statistics rows for legacy tests/memory metrics
This CL allows us to achieve test hierarchy parity with the /add_point
pipeline. In particular, for most types of tests, we want to create the same
suffixed statistics rows that /add_point does from chart JSON. There are two
exceptions to this: memory metrics, for which we want to only produce _avg,
and legacy tests, for which we don't want to produce these suffixed rows at
all.

Fixes #4201.

Bug: catapult:#4203, catapult:#4201
Change-Id: Ic258a43bc1482a205b4cac4fa3f306f55488b4c4