[pinpoint] Decrease questionable threshold as repeats increase.
The more repeats there are, the more strict the threshold becomes for doing more
repeats. This will help reduce the number of extraneous repeats and reduce the
bot load.

This will also help us scale for flaky functional bisects, which may need dozens
or hundreds of repeats to gain enough confidence for a failure that happens
rarely (<20% of the time).

By setting the threshold using representative samples, we can tune the threshold
to samples of a specific noise level.

Change-Id: I2ffd22da638473a4959774fb6f861a6d30bef182