[Telemetry] Move browser and profile directories to base chrome backend
As the interface is now uniform for all of desktop, cros, and android,
access to the browser and profile directories can now be handled by
the base ChromeBrowserBackend class.

Also tidy up the args for all backends to be consistent and follow
the order: platform_backend, browser_options, browser_directory,
profile_directory, [... platform specific args ...].

Take the change to move the check of "no exceptions to load when
extensions are not supported" to the base class.

Bug: chromium:805486
Change-Id: Ic76c8c8566d542f8a7f8a2eb9ab9b5d92807cfa4