[Telemetry] Install CrOS browser extensions during environment set up
Copying the extensions to local paths on the device (and clearing them
up when finishing) will now be managed by the possible browser
environment set up (and tear down).

This will allow, on a follow up CL, to finalize the computation of
startup args for chrome, which depend on knowing the local paths for
such extensions.

Bug: chromium:787834
Change-Id: I4f518b41c361f2a75d56b8462078afeee5f57087