Update patch set 3
Patch Set 3:

(3 comments)

/cc tdresser@ FYI. 

Thanks a lot for this patch! The approach is looking good in general. 

> I thought it is better to do them in a followup cl.

Unfortunately, not fixing those tests in this CL means we will have to disable those tests until they're fixed. We want to avoid that. How about this CL sequencing?  
CL 1: Implement all this in load expectations, but leave the current implementation in loading_metric. You can add a comment in loading_metric with a bug saying the code there is duplicate and is scheduled to be removed soon.
CL 2: Fix the tests to switch over to exclusively using load expectations.
CL 3: Delete the duplicate code in loading_metric.

Patch-set: 3
Reviewer: Gerrit User 1177266 <1177266@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>