Update load expectations to use time to interactive
Load expectations uses first contentful paint events to determine
load expectations. This is not representative of loading and it
is better to use time to interactive instead. Time to interactive
is already used by the loading metric. This cl moves the computation
of time to interactive to load expectation and updates loading metric
to use load expectations to obtain loading related metrics.


Bug: chromium:796109
Change-Id: Ic75d8ca78a66103e101b66201b0fa49a76f38d8b