Fix NamedDeviceTemporaryDirectory leak with contextlib_ext.Optional
NamedDeviceTemporaryDirectory creates its managed directory in its
__init__ method, but deletes it in __exit__. Since
contextlib_ext.Optional unconditionally calls __init__ but conditionally
calls __exit__, the directory will be leaked if its condition is False.
This fix moves the creation of the directory to __enter__ so that the
directory is only created if it will also be destroyed.

BUG=809192

Change-Id: Ib1dced2fecf32ab343e8c4f53762f1797811a541