[Telemetry] Add PageTestResults.Fail hook
Rather than adding FailureValues, we want users of PageTestResults to use a
hook that abstracts this logic instead. This is so that we can migrate
FailureValue out of the value system and make it easier to replace
Telemetry's value system with the Catapult histogram system.

This CL adds that hook, PageTestResults.Fail, which is currently implemented
in terms of FailureValue.

Bug: catapult:#4229
Change-Id: I16ee51c34985278d02ba59927735fa5c4f8b574f