Always report FPS in webvrMetric
Ensures that the webvr_fps histogram created by the webvrMetric metric
is always reported. If no actual tracing data is available, then 0 is
reported.

This is to ensure that a failure to submit frames shows up as a
regression on the perf dashboard instead of just not reporting any
new data.

Bug: chromium:810218
Change-Id: Ib2c9150175f0cbc9c9a9d1ade4effdb680b5d2f5