[Devil] Manage SELinux policy within FlagChanger object
This used to be handled by the CustomCommandLineFlags context
manager wrapper. However, some clients would like to use the FlagChanger
directly, while not worrying about having to manually set/reset the
policy as needed.

With this change calls to any of the flag "setting" methods (Replace-,
Add-, RemoveFlags) will also SetPermissive; while the last call to
Restore (when clearing the stack and returning to the original flags
on the device) will also ResetEnforce.

TBR=jbudorick@chromium.org
Bug: chromium:787834
Change-Id: I64345c6d16adae80e075d9fd69005aa26082ad19