Pinpoint - Add options when querying for job data.
This adds an 'o' parameter that lets you specify specify job
properties you want returned. This also adds a 'tags' field to jobs,
that can be set on creation. This is a dict of string/string pairs
that can hold metadata (from the dashboard).

Options right now are:
STATE: The full job state.
DIFFERENCES: A list of the changes where differences were detected.
TAGS: The 'tags' metadata from the job entity.

Bug: catapult:#4217

Change-Id: I4212d846e411c3528adee3b394745cb82cfb0a50