Pinpoint - Add a re-run button.
This replaces the existing floating action button which opens a perf
try dialog. Now it expands out into several buttons, one of which is
the new re-run button.

Eventually we probably want to let people customize all the options
for re-running a job, but for now I just made it straight up send the
same params to create a new job.

Bug: catapult:#4210
Change-Id: Ie550e8e3d4fa0d51d0bed1e6f80f8b2ee1211f61