[Telemetry] Move cache clearing code to possible_browser
As a first step towards shifting the ownership of cache clearing
decisions to the shared state, this CL:

- provides a new possible_browser.FlushOsPageCaches() method
- replaces browser.ClearCaches() with a new private method
  possible_browser._ClearCachesOnStart()

There should be no behavior change on cache clearing at the moment.

On follow up CLs _ClearCachesOnStart should be removed when the
shared state code calls individual cache clearing methods as needed.

Bug: chromium:811244
Change-Id: Ie9c274b2211eb1e1eff5aa5e074496a64df1aac3