[pinpoint] Auto-apply fixes for failing commit ranges.
This CL doesn't include the ability to automatically detect failing ranges, so
right now the best way to add a fix to the database is using the dev console.

The fix patch is applied whenever an Attempt is created; the Changes at the top
level of the JobState don't include the patch, to make UI and Midpoint easier.
See the bug for a lengthier discussion on when the fix is applied.

Fix lookup is O(1) - we do a single ndb key lookup for each commit.

Bug: catapult:#3516
Change-Id: I41113bf78a5a674e7c0b473ca033cc79f9ee6de2