[Telemetry] Move apk finding code to Android backend settings
A first step towards encapsulating the logic to find and install apks
for individual Android browser variants, this CL introduces a
backend_settings.FindLocalApk() method.

As a side effect of this change the logic of the corresponding browser
finder becomes cleaner and now, even if --chrome-root is not given,
allows to select any known browser which is already installed on the
device.

We also take the chance to make _HaveLocalAPK and
_HaveWebViewEmbedderAPK private, as the logic to handle those will
be eventually moved to the banckend_settings as well.

Bug: chromium:815133
Change-Id: I92315d5e169ed308ea685213203244ba4812c186