[soundwave] Add integration test for FetchAlertsData
Add an integration test for the command to FetchAlertsData. It uses a
mock dashboard_communicator to return fake alert data, runs the command,
and checks that the alerts stored on a (temporary) database match what
we expect.

For better organization we also move most of the code that used to be on
the bin/soundwave script to its own soundwave.commands module.

Bug: catapult:#4293
Change-Id: I1118df4eb7b04536940dec86df318d44af5bd959