Tracing: CPU time computation clean up
tr.model.Thread.getCpuStatsForRange is not used anywhere. Instead, there is a
code for computing CPU time of a thread in a range that is duplicated in two
places, for some reason, and is more useful.  This CL deletes
tr.model.Thread.getCpuStatsForRange and replaces it with the more interesting
version. This is in preparation for migrating thread times metrics to TBMv2.

Bug: chromium:627461
Change-Id: I8bfe6813dd8e0c69c352f61fa61f4620e53449d2