[pinpoint] Generalize "configuration" to be a set of argument presets.
Previously, the individual FindIsolate and RunTest Quests would look for their
special parameters in the `bot_configurations` map. Instead, this treats
"configuration" as a special argument; if specified, it adds the entire contents
of the config in the `bot_configurations` map to the current parameter list. In
this way, "configuration" is treated as a set of argument presets.

Bug: catapult:#4336
Change-Id: I350f23b0b3cadf60c57abbee86c7167750e63103