Update patch set 2
Patch Set 2: Code-Review-1

Sorry, but no, I don't like in particular passing this down as an argument to the browser constructor.

Instead, the client (a shared state maybe) should create the browser, then use some API on it to check whether gpu rendering is enabled and react to that information (raise an exception, or whatever the client wants to do with it).

Is this something we want for a specific benchmark? Or all benchmarks on a specific platform? In which case were we thinking about using this flag?

Patch-set: 2
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1