Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > Sorry, but no, I don't like in particular passing this down as an argument to the browser constructor.
> > > 
> > > Instead, the client (a shared state maybe) should create the browser, then use some API on it to check whether gpu rendering is enabled and react to that information (raise an exception, or whatever the client wants to do with it).
> > > 
> > > Is this something we want for a specific benchmark? Or all benchmarks on a specific platform? In which case were we thinking about using this flag?
> > 
> > we want this for all benchmarks & gpu correctness test, that's why I put it in the browser object.
> 
> I understand why it's awkward to put this in browser constructor. How about put this logic in browser_finders code?
> 
> Basically they are are assert statement about correctness statement of the created browser.

So we want to assert this for *all* browsers created? Why is it then an option at all?

It would be nice to put in the browser finder, but it sounds like you need to actually start the browser to check if gpu rendering is working? (and finding the browser should not start it.)

If you really want to have this as an option, and really want to keep it deeper than the shared state, then it _could_ belong to the possible_browser, since the possible_browser is the one taking the "finder_options" blob and preparing a browser according to those. I would like to keep the Browser object itself as agnostic as possible from those options.

Patch-set: 2
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>