[pinpoint] Show "my jobs" only.
When you are logged in, the main page will only jobs you kicked off (from both
perf dashboard and Pinpoint). If you are logged out, it will still show all
jobs. This behavior is similar to Gerrit and Monorail.

For Pinpoint triaging, it may still be useful to list all jobs. The workaround
right now is to log out while triaging. In the future, we want to use /stats
for this, as the central place for all things related to monitoring.

Bug: catapult:#4257
Change-Id: I8ce04d23899a069988e55f6e57dff160f181e373