Add --enable-automation flag in GetFromBrowserOptions()
crrev.com/c/956654 introduced the bad flags warning of "stability and
security will suffer" for Android.
This change caused parformance regressions because a bad flag
"--ignore-certificate-errors-spki-list" is used for perfomance testing.
(crbug.com/822258#c13)
So crrev.com/c/965727 introduces the check of "--enable-automation" flag in ShowBadFlagsPrompt().
(This check already exists in startup_browser_creator_impl.cc for Desktop not
to show the infobar during performance tests.)

This CL adds the flag to GetFromBrowserOptions().

Bug: chromium:822258
Change-Id: I8a07c88960678877ee1add1c0d29ad323611a816