Update patch set 2
Patch Set 2:

Please note that we need to do _no_ merging of zero-story histograms. This means that they should not even be passed into the by-repeats merge. This turns out to be surprisingly messy, so let's talk about it tomorrow.

The main problem is that we don't have a good way of porting a histogram from one set to another inclusive of its shared diagnostics. In JS, we make a distinction between addSharedDiagnostic, which only adds a shared diagnostic to the map, and addSharedDiagnosticToAllHistograms, which adds the diagnostic to the map AND adds references to it to all histograms in the set. At a minimum, this would make what we have to do here more elegant. There is an inefficient way to do it, though:

all_histograms = histogram_set.HistogramSet()
all_histograms.ImportDicts(input_path)

no_merge = []
to_merge = []

for h in all_histograms:
  hs = histogram_set.HistogramSet([h])
  for d in h.diagnostics:
    if d.has_guid:
      hs.AddSharedDiagnostic(d)
  if reserved_infos.STORIES.name in h.diagnostics:
    to_merge.append(hs)
  else:
    no_merge.append(hs)

all_to_merge = histogram_set.HistogramSet()
for hs in to_merge:
  all_to_merge.ImportDicts(hs.AsDicts())
all_to_merge.DeduplicateDiagnostics()

# Write all_to_merge.AsDicts() to a file to_merge_dicts
# Do both merging steps on to_merge_dicts
# Create the HistogramSet final_histograms and import both sets of dicts as before

for hs in no_merge:
  final_histograms.ImportDicts(hs.AsDicts())
final_histograms.DeduplicateDiagnostics()

# etc.

Given AddSharedDiagnostic as distinct from AddSharedDiagnosticToAllHistograms, we could skip this clunky HistogramSet-per-Histogram logic and a lot of json.loads and json.dumps.

Patch-set: 2
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>