Update patch set 3
Patch Set 3:

Sorry I reverted it because I remembered time breakdowns can still be negative because of the overlapping slice issue, and it may break things if we throw a hard error. 

Following the discussion at https://github.com/catapult-project/catapult/issues/4267, it seems Ethan was looking for a more generic solution, and a lot of the trouble was coming from memory metric breakdowns too. Ben if you think this one-off fix in loading_metric is still useful I'll reland it with just a "!== 0" check. Otherwise (esp since it looks like they already have worked around the problem for the moment by purging all breakdowns) we can wait till we have a more general fix.

Patch-set: 3