Pinpoint - Add closing </p> tag to new-job-fab
Dave noticed this wasn't submitting properly anymore. Digging in deeper
we found that submit() was bailing out early, this._form isn't
initialized (which happens on attached()), and the iron-form wasn't
finding it's form nodes. Weirdly, adding this tag to the open <p>
fixed the issue. Not clear why, but let's get the form back up and
running at least.

Change-Id: Id60ddf74b18c20d2ad1bf55180e0560d51f43d4b