[pinpoint] Separate RunTest Quest into Telemetry and GTest Quests.
Separating the Telemetry and GTest code paths solves the --results-label
awkwardness, since we only need to worry about it in the Telemetry code path.

We also no longer need to worry about the target in the Quest, since that was
only used to distinguish Telemetry from GTest.

It also makes future changes easier, since the parameter handling and execution
logic are split into separate files, for both the actual code and the unit
tests.

Bug: catapult:#3998
Change-Id: I938045b78872d24422a7c453e08b95d11a06672e