Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > > Patch Set 2:
> > > > > 
> > > > > Can this issue wait for Perfetto to revamp the UI later this year?
> > > > > 
> > > > > Alternatively, can you pull the Sinebow color generator changes from this CL? It works similarly to your hue-wheel math, but accounts for some human perceptual quirks and provides a more general API.
> > > > > https://chromium-review.googlesource.com/c/catapult/+/938391/152/tracing/tracing/base/sinebow_color_generator.html
> > > > > 
> > > > > Also, please link to screenshots comparing the old colors to the new colors both in a swatch and in context of a trace. The existing colors were hand-picked a while ago, so we should probably look at the new colors closely (on different screens, in context, etc) before switching away from them.
> > > > 
> > > > Thanks for the review. I'll prepare a screenshot soon. While the existing colors were habd-picked, I'm certain the new ones are better wrt readability since the old ones include some really dark ones.
> > > > 
> > > > While I'm happy there is going to a revamp, this change is fairly straightforward, so I was hoping this can land first, as I think it provides a quick net win for usability.
> > > 
> > > Here's the new color palette: https://bit.ly/2Gv9n8T
> > > 
> > > The old one is on the top row, the new one on the bottom.
> > 
> > In case that link doesn't work... https://i.imgur.com/7xQgPuW.png
> 
> Any follow up on this? I'm fine with abandoning this CL if an UI overhaul is coming soon. If not, this CL is a quick way to make things look nicer in the meantime.

A few followups:
Ben: I checked with oysteine@ and perfetto is not going to affect that part of the code and need to be maintained for at least another year as is.

Concerning the color palette itself:
I found it problematic on many traces. The biggest issue is bright colors blend together.
This is very clear when we have green and cyan events, when zoomed out it would be missed out as one big event.
Similarly, the white strips between yellow events are non-visible since the yellow is so pale.
Examples:
https://imgur.com/dkefdsn
https://imgur.com/g9lJ4Sq

In the second one, you can also see the purple bleeding into the red.

Without using exact science here, I do not find the new palette (with the current configuration) to be a good fit for trace-viewer at the moment.

If Ben or Ethan disagree, then I have no objection to landing it (after fixing the tests).

Patch-set: 2
Reviewer: Gerrit User 1115914 <1115914@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>