Update patch set 3
Patch Set 3:

I have done further research.
I agree that the labels are not very visible on the current color palette.

But the strength of the current palette is that it tends to pick very different colors for similar strings (labels). This is due to 2 things: first the colors are added in a pseudo-random way, allowing the (second) the hashing to pick different colors for similar strings.

The new palette has better contrast, but tends to pick similar color for similar strings.

For zoomed in traces, the new palette is clearly better, but when zoomed out it is worse.
This also might be fixed by changing how we draw the fast rectangles (when we skip some rectangles when zoomed out).

Adding some side by side comparisons:
https://imgur.com/E7uUpYl (zoomed in, new palette is better)
https://imgur.com/Qtm7NZb (zoomed out, colors blending)
https://imgur.com/QkgfTfP (mixed zoom, shows strengths and weaknesses)

Patch-set: 3
Reviewer: Gerrit User 1115914 <1115914@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>