Add a regression test for browser-side fling
The fling gesture is now handled on browser. But we discovered a bug
for that. When there is a in-flght fling gesture and a second fling comes,
we suppress the ScrollBegin for the second fling gesture, and thus the
touch-action filter will not set the correct touch action, which could
allow the second fling that should be blocked.

This CL adds a new test for this case. We create a div with pan-x, and
start swiping on that div along the x-direction, before it stopped flinging,
we swipe the same div on y-direction, and it should be allowed.

Bug: 828436
Change-Id: I29c62fa0f0635eb71178a874ec142cd5126a999e