Implement Browser Resize Api in action_runner
We would like to write stories which can be used to measure the performance of
resizing the browser.

This adds a new function to action_runner which can trigger resizes of the
browser window. With optional parameters to do this over a series of steps. Thus
allowing the simulation of a user performing a drag to resize the browser.

Bug: catapult:#
Change-Id: Ic23b3936b64ac5e11c642550e5d135a67771d9f4