HistogramSet - Merge by test paths in add_reserved_diagnostics.
Refactored add_reserved_diagnostics to compute the dashboard test paths
and then merge using that as the key. This is more straightforward
than previous implementations, at the cost of some efficiency.

This is largely to support summaries at both the metric and tir_label
level, so is_summary has been modified to be a set of names which
represent what was merged to create this summary statistic. The
dashboard picks that up so that it can properly decide how to create
the test path.

Bug: catapult:#4391

Change-Id: I1a136e581ba5b1e0706d499cd2cd0a58a2ec78b1