[pinpoint] Use number of attempts to determine significance threshold.
Previously we used the number of values instead of the number of attempts. This
is a problem if early test runs succeed, but all subsequent test runs fail.
Because the number of values and the threshold both remain the same, the
comparison keeps giving "need more data", and we just keep on retrying forever.

Using the number of attempts instead penalizes retries.

Bug: catapult:#4393
Change-Id: I1b4078c0ffa55e7ef74215647d620f5d00d0cac1