Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> OK, made some tweaks, added some tests, and ran it using Max's described method locally. Things seem ok to do a first run from our side.
> 
> We have quite a few ideas on variants on this metric, but we thought it might be better to get some initial data on the correlations with FID/TTI before investing a lot more time into it, if that sounds good to you folks as well? If it's incredibly difficult to rerun with different variants and you'd much prefer to do it all in one go, let us know.
> 
> Thanks again for the assist! :)

By "Rerun" I guess you mean rerun the metric computation of variants on the same trace data. Currently we use Chrome Trace Processor to do the rerun, but the environment setup is pretty tricky, and ned has just proposed an alternative to do it:
https://docs.google.com/document/d/1MY95ULhEuKFznBQpF60_uhdhRco5bWzUVfVTp0M2hDw/edit?ts=5accff78#heading=h.y1yx1wbo1e9y

But as we've already setup-ed the environment for our previous experiments, it should be less frictions for us to do it:)

Let's do an initial run first to collect the traces and get some initial data. After you get a sense with the data, you can add variants. Try to come up with as many variants as possible and put them into different histogram. And we can rerun them with CTP.

Patch-set: 6
Reviewer: Gerrit User 1215729 <1215729@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>