Use top-level scheduler tasks in EQT metric.
The EQT metric is currently uses message loop tasks, but input events
are blocked on scheduler tasks.

This patch extracts the top level scheduler tasks finding function from
the TTI metric and uses it in both EQT and TTI metrics.

The finding function now does only one pass over the slices and preserves
the order of slices in the result.

Bug: chromium:832651
Change-Id: I0127563b1ba9cf318b5c6b066b9a73235d1b389a