Update patch set 5
Patch Set 5:

OK, I dug into the behavior of the ps command on oreo, and it's pretty ugly...

The flag to show threads on oreo is '-T', but it's quite broken.  This:

$ ps -p 1234 -T

... will only print one line, for the process with PID 1234; it doesn't show the threads.  Adding '-e' to the command doesn't help.  You can see all threads associated with all processes via:

$ ps -e -T

The only way I could figure out to get a listing of just the threads associated with a particular PID is:

$ ps -e -T | grep '^[[:graph:]]\{1,\}[[:blank:]]\{1,\}PID[[:blank:]]\{1,\}'

... where 'PID' in the regex pattern is replaced by the numeric pid of interest.

That is do-able, but if the signature of ListProcesses becomes:

def ListProcesses(self, process_name_pattern=None, pid=None, list_threads=False)

... then it gets pretty ugly when you combine process_name_pattern='foo' and list_threads=True.  You would have to first:

$ ps -e | grep -F PROCESS_NAME_PATTERN

... then parse the output for pid's in the result, and run a second command:

$ ps -e -T | grep ''^[[:graph:]]\{1,\}[[:blank:]]\{1,\}(PID1|PID2|PID3|...)[[:blank:]]\{1,\}'

... with PID1, PID2, PID3, etc., replaced by the pid's parsed from the first ps command's output.

So, all things considered, I have two proposals:

1. Go back to my initial proposal: don't change ListProcesses, and add:

def ListThreadsForProcess(self, pid)

That will still need to used a piped shell command with the wacky regex, but it's not too terrible.

2. Drop this CL entirely and push the wacky logic, version sniffing, and piped shell commands down into my telemetry code patch, which is the only call site.

WDYT?

Patch-set: 5
Reviewer: Gerrit User 1001706 <1001706@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>